+++ Changes.md
@@ -15,6 +15,7 @@ Improvements
 - 3Delight : Added NSI screen static sampling pattern option (`dl:staticsamplingpattern`).
 - GraphEditor : The source node for any location can be located by dragging the location into the GraphEditor from the Viewer or HierarchyView.
 - GraphComponent : Node and Plug names may now start with a numeric digit.
+- 3Delight : Added support for `layerName` parameter in output definitions.
 
 Fixes
 -----
@@ -30,6 +31,11 @@ Fixes
   - Fixed errors when dragging from another application onto widgets that didn't expect it.
 - Gadget : Fixed access to `DragDropEvent.sourceWidget` and `DragDropEvent.destinationWidget` from Python slots connected to a Gadget's DragDropSignals.
 - GraphGadget : Fixed unwanted highlighting of nodes when custom drag & drop handlers were active.
+
+Breaking Changes
+----------------
+
+- 3Delight : Changed default layer names for outputs.
 
 1.4.0.0b6 (relative to 1.4.0.0b5)
 =========